Fix build on Windows
wslua_struct.c(431) : error C2220: warning treated as error - no 'object' file generated
wslua_struct.c(431) : warning C4244: '=' : conversion from 'lua_Number' to 'size_t', possible loss of data

Reviewed-on: https://code.wireshark.org/review/213
Tested-by: Alexis La Goutte <alexis.lagoutte@gmail.com>

+++ epan/wslua/wslua_struct.c
@@ -428,7 +428,7 @@ WSLUA_CONSTRUCTOR Struct_unpack (lua_State *L) {
         if (size == 0) {
           if (!lua_isnumber(L, -1))
             luaL_error(L, "format `c0' needs a previous size");
-          size = lua_tonumber(L, -1);
+          size = (guint32)lua_tonumber(L, -1);
           lua_pop(L, 1);
           luaL_argcheck(L, pos+size <= ld, 2, "data string too short");
         }
